+++ PVEFrame2.lua
@@ -31,23 +31,8 @@ local function createCategoryButtons(categoryID, type, rootDescription)
 
 			if(type == "search") then
 				openSearchPanel(categoryID, filters)
-
-				miog.ProgressPanel:SetShown(categoryID == 2 or categoryID == 3)
-
-				miog.ProgressPanel.MythicPlus:SetShown(categoryID == 2)
-				miog.ProgressPanel.Raids:SetShown(categoryID == 3 and filters == Enum.LFGListFilter.Recommended)
-
-				miog.ProgressPanel.Background:SetTexture(miog.ACTIVITY_BACKGROUNDS[LFGListFrame.SearchPanel.categoryID])
-
-				local playerName, realm = miog.createSplitName(UnitFullName("player"))
-				
-				miog.ProgressPanel:Flush()
-				miog.ProgressPanel:SetPlayerData(playerName, realm)
-				miog.ProgressPanel:ApplyFillData()
 				
 			else
-				miog.ProgressPanel:Hide()
-
 				LFGListFrame.CategorySelection.selectedCategory = categoryID
 				LFGListFrame.CategorySelection.selectedFilters = filters